scripts/alert-monitoring/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
slack_sdk>=3.21.0
pyyaml>=6.0.1
